requirements.txt
openai
python-dotenv
pydantic
pyyaml
datasets<3.0.0
transformers
